Some fixes to instructions to get working on x86_64 systems

+++ benchmark/tfs_grpc_client.py
@@ -37,7 +37,7 @@ stub = prediction_service_pb2_grpc.PredictionServiceStub(channel)
 isd = lambda a, b, c : {'inputs': a, 'shape': b, 'dtype': c}
 
 models = {
-          'small_lstm': isd('inputs', (args.batch, 8, 48), np.float64),
+          'small_lstm': isd('inputs', (args.batch, 8, 48), np.float32),
           'medium_cnn': isd('inputs', (args.batch, 101, 82, 9), np.float32),
           'large_tcnn': isd('inputs', (args.batch, 3, 101, 82, 9), np.float32),
           'swmodel': isd('dense_input', (args.batch, 3778), np.float32),
